requirements.txt
Flask==0.10.1
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
argh==0.25.0
argparse==1.2.1
click==3.1
colorama==0.3.1
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
gfm==0.0.3
greenlet==0.4.3
itsdangerous==0.24
lockfile==0.9.1
lxml==3.3.5
nose==1.3.3
pathtools==0.1.2
python-daemon==1.6.1
watchdog==0.8.1
wsgiref==0.1.2
html2text==2014.7.3

# For now, use my fork which disables gevent monkey patching.
# It was interrupting `watchdog`, and the lib seems to work fine without it.
git+git://github.com/ftzeng/Flask-SocketIO

# The py-gfm package is not in pypi, so install from git.
git+git://github.com/dart-lang/py-gfm
